DCA/requirements.txt
# The following packages, commented with "#", are recommended for installation in the listed order.

networkx==3.0
numpy==1.22.3
pandas==1.4.2
pydot==1.4.2
scikit-learn==1.2.1
scipy==1.10.0
# torch==1.11.0
# torch-scatter==2.0.9
# torch-sparse==0.6.15
# torch-cluster==1.6.0
# torch-geometric==2.1.0.post
pytorch-warmup==0.1.1